test (automated command)

+++ 0x02-functions_nested_loops/0-putchar.c
@@ -11,7 +11,7 @@
 int main(void)
 {
 	char string[] = "_putchar";
-	int i = 0;
+	long int i = 0;
 
 	for (; i < sizeof(string); i++)
 	{
